requirements.txt
flask==2.3.2
flask-session==0.5.0
qrcode==7.3.1
python-dotenv==0.20.0
Flask-API==3.1
